+++ ands/__init__.py
@@ -1,4 +1,4 @@
 from .models import Evidence, ProbeResult, ScanReport
 from .scanner import openapi_hints, pick_probe_paths, infer_ands, map_to_regulations, analyze_probe_status, create_bundle
-from .validator import verify_declaration_signature
-from .utils import get_session, safe_request, normalize_base_url, check_tls_integrity, logger
+from .validator import verify_declaration_signature, validate_declaration
+from .utils import get_session, safe_request, normalize_base_url, check_tls_integrity, get_supported_versions, logger